+++ shoop/xtheme/static_src/injection/index.js
@@ -48,6 +48,7 @@ function setSidebarVisibility(visible) {
 
 function openPlaceholderEditor(domElement) {
     const placeholderName = domElement.dataset.xtPlaceholderName;
+    const globalType = domElement.dataset.xtGlobalType;
     if (!placeholderName) {
         return;
     }
@@ -58,6 +59,7 @@ function openPlaceholderEditor(domElement) {
         view: viewName,
         theme: window.XthemeEditorConfig.themeIdentifier,
         ph: placeholderName,
+        "global_type": globalType,
 
         // TODO: Hopefully we won't get any problems with too-long query strings (2048 is the maximum for IE):
         "default_config": defaultConfigJSON