+++ doc/changelog.rst
@@ -72,6 +72,7 @@ Xtheme
 ~~~~~~
 
 - Enhance default text plugin editor to remarkable markdown editor
+- Add support for global/multi-view placeholders
 - Add generic snippets plugin for doing simple integrations
 - Add a plugin for displaying category links on shop front
 - Add a linkable image plugin